requirements.txt
streamlit
ultralytics>=8.3.0
easyocr
opencv-python-headless
pillow
numpy
pypdfium2